+++ packages/example/src/services/complex/adapters/mod.rs
@@ -1,2 +1,0 @@
-pub mod complex_repository_adapter;
-pub mod complex_service_adapter;