+++ packages/example/src/services/complex/ports/mod.rs
@@ -1,2 +1,0 @@
-pub mod complex_repository_port;
-pub mod complex_service_port;