+++ packages/example/src/services/mod.rs
@@ -1,2 +1,3 @@
-pub mod complex;
+pub mod mutex;
+pub mod rw_lock;
 pub mod user_service;