+++ packages/example/src/services/mutex/adapters/mod.rs
@@ -1,0 +1,2 @@
+pub mod mutex_repository_adapter;
+pub mod mutex_service_adapter;