+++ packages/example/src/services/mutex/commands/add_element_command.rs
@@ -1,10 +1,10 @@
 use std::{error::Error, sync::Arc};
 
 use async_trait::async_trait;
-use kti_cqrs_rs::common::handler::CommandHandler;
+use kti_cqrs_rs::common::handler::command_handler::CommandHandler;
 use tokio::sync::Mutex;
 
-use crate::services::complex::contexts::complex_context::ComplexContext;
+use crate::services::mutex::contexts::mutex_context::MutexContext;
 
 pub struct Command {
   pub element: i32,
@@ -12,13 +12,13 @@ pub struct Command {
 
 #[async_trait]
 impl CommandHandler for Command {
-  type Context = ComplexContext;
+  type Context = Arc<Mutex<MutexContext>>;
   type Output = Result<(), Box<dyn Error>>;
 
-  async fn execute(&self, context: Arc<Mutex<Self::Context>>) -> Self::Output {
+  async fn execute(&self, context: Self::Context) -> Self::Output {
     let ctx = context.lock().await;
     let command_repository = ctx.get_command_repository();
 
-    command_repository.remove_element(self.element).await
+    command_repository.add_element(self.element).await
   }
 }