+++ packages/example/src/services/rw_lock/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod add_element_command;
+pub mod remove_element_command;