+++ packages/example/src/services/rw_lock/ports/mod.rs
@@ -1,0 +1,2 @@
+pub mod rw_lock_repository_port;
+pub mod rw_lock_service_port;