+++ packages/example/src/services/user_service/get_user_by_name_query.rs
@@ -1,7 +1,7 @@
 use std::{error::Error, sync::Arc};
 
 use async_trait::async_trait;
-use kti_cqrs_rs::common::handler::QueryHandler;
+use kti_cqrs_rs::common::handler::query_handler::QueryHandler;
 use tokio::sync::Mutex;
 
 use super::user_service::{User, UserService};
@@ -20,10 +20,10 @@ impl GetUserByNameQuery {
 
 #[async_trait]
 impl QueryHandler for GetUserByNameQuery {
-  type Context = UserService;
+  type Context = Arc<Mutex<UserService>>;
   type Output = Result<Option<User>, Box<dyn Error>>;
 
-  async fn execute(&self, context: Arc<Mutex<Self::Context>>) -> Self::Output {
+  async fn execute(&self, context: Self::Context) -> Self::Output {
     let user_service = context.lock().await;
 
     user_service.get_user_by_name(&self.name)