+++ packages/kti_cqrs_rs/src/common/context/mod.rs
@@ -1,3 +1,1 @@
-mod context;
-
-pub use context::Context;
+pub mod context;