+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.2.0 BREAKING CHANGES
+* Modify context for more flexible control of types
+* Add example for `tokio::sync::Mutex`
+* Add example for `tokio::sync::RwLock`
+
 ## Version 0.1.0 BREAKING CHANGES
 * Replace the `std::sync::Mutex` to `tokio::sync::Mutex`
 * Add new example for complex async tests